+++ internal/dinosaur/pkg/cmd/centrals/cmd.go
@@ -1,6 +1,6 @@
 // Package central contains commands for interacting with central logic of the service directly instead of through the
 // REST API exposed via the serve command.
-package central
+package centrals
 
 import (
 	"github.com/spf13/cobra"
@@ -10,10 +10,11 @@ const (
 	apiErrorMsg = "%s Central failed: To fix this ensure you are authenticated, fleet-manager endpoint is configured and reachable. Status Code: %s."
 )
 
-// NewCentralCommand ...
-func NewCentralCommand() *cobra.Command {
+// NewCentralsCommand ...
+func NewCentralsCommand() *cobra.Command {
 	cmd := &cobra.Command{
-		Use:              "central",
+		Use:              "centrals",
+		Aliases:          []string{"central"},
 		Short:            "Perform central CRUD actions directly",
 		Long:             "Perform central CRUD actions directly.",
 		PersistentPreRun: func(cmd *cobra.Command, args []string) {},
