+++ internal/dinosaur/pkg/cmd/centrals/delete.go
@@ -1,4 +1,4 @@
-package central
+package centrals
 
 import (
 	"fmt"
