+++ internal/dinosaur/pkg/cmd/centrals/flags.go
@@ -1,4 +1,4 @@
-package central
+package centrals
 
 // FlagID ...
 const (
